Move chown from fs.rs to ops/fs.rs
